fleetbench res

+++ src/scripts/setups/setup_utils.sh
@@ -124,22 +124,26 @@ function setup_pkb() {
 function setup_multichase() {
     echo $(setup_echo "Multichase Requirements")
 
-    sudo apt-get install -y make gcc numactl
+    if [ $(setup_pkgExists gcc) -eq 0 ]; then
+        sudo apt-get install -y make gcc numactl
+    fi
 }
 
 function setup_fleetbench() {
     echo $(setup_echo "Fleetbench Requirements")
 
-    if [ "$ARCH" == "intel" ]; then
-        curl -fsSL https://bazel.build/bazel-release.pub.gpg | gpg --dearmor >bazel-archive-keyring.gpg
-        sudo mv bazel-archive-keyring.gpg /usr/share/keyrings
-        echo "deb [arch=amd64 signed-by=/usr/share/keyrings/bazel-archive-keyring.gpg] https://storage.googleapis.com/bazel-apt stable jdk1.8" | sudo tee /etc/apt/sources.list.d/bazel.list
-        sudo apt-get update && sudo apt-get install -y bazel
-    elif [ "$ARCH" == "arm" ]; then
-        BAZEL_LATEST_VERSION=$(curl -s https://api.github.com/repos/bazelbuild/bazel/releases/latest | grep tag_name | cut -d '"' -f 4)
-        wget https://github.com/bazelbuild/bazel/releases/download/$BAZEL_LATEST_VERSION/bazel-$BAZEL_LATEST_VERSION-linux-arm64
-        chmod +x bazel-$BAZEL_LATEST_VERSION-linux-arm64
-        sudo mv bazel-$BAZEL_LATEST_VERSION-linux-arm64 /usr/local/bin/bazel
+    if [ $(setup_pkgExists bazel) -eq 0 ]; then
+        if [ "$ARCH" == "intel" ]; then
+            curl -fsSL https://bazel.build/bazel-release.pub.gpg | gpg --dearmor >bazel-archive-keyring.gpg
+            sudo mv bazel-archive-keyring.gpg /usr/share/keyrings
+            echo "deb [arch=amd64 signed-by=/usr/share/keyrings/bazel-archive-keyring.gpg] https://storage.googleapis.com/bazel-apt stable jdk1.8" | sudo tee /etc/apt/sources.list.d/bazel.list
+            sudo apt-get update && sudo apt-get install -y bazel
+        elif [ "$ARCH" == "arm" ]; then
+            BAZEL_LATEST_VERSION=$(curl -s https://api.github.com/repos/bazelbuild/bazel/releases/latest | grep tag_name | cut -d '"' -f 4)
+            wget https://github.com/bazelbuild/bazel/releases/download/$BAZEL_LATEST_VERSION/bazel-$BAZEL_LATEST_VERSION-linux-arm64
+            chmod +x bazel-$BAZEL_LATEST_VERSION-linux-arm64
+            sudo mv bazel-$BAZEL_LATEST_VERSION-linux-arm64 /usr/local/bin/bazel
+        fi
     fi
 }
